refactor: move ipc handlers to modules
